+++ registry/node_test.go
@@ -2,6 +2,7 @@ package registry
 
 import (
 	"context"
+	"net/http"
 	"strings"
 	"testing"
 	"time"
@@ -9,7 +10,8 @@ import (
 	dc "github.com/bilibili/discovery/conf"
 	"github.com/bilibili/discovery/model"
 	"github.com/bilibili/kratos/pkg/ecode"
-	http "github.com/bilibili/kratos/pkg/net/http/blademaster"
+	bm "github.com/bilibili/kratos/pkg/net/http/blademaster"
+	"github.com/bilibili/kratos/pkg/net/http/blademaster/binding"
 	xtime "github.com/bilibili/kratos/pkg/time"
 
 	. "github.com/smartystreets/goconvey/convey"
@@ -20,14 +22,13 @@ var config = newConfig()
 
 func newConfig() *dc.Config {
 	c := &dc.Config{
-		HTTPClient: &http.ClientConfig{
+		HTTPClient: &bm.ClientConfig{
 			Timeout:   xtime.Duration(time.Second * 30),
 			Dial:      xtime.Duration(time.Second),
 			KeepAlive: xtime.Duration(time.Second * 30),
 		},
-		HTTPServer: &http.ServerConfig{Addr: "127.0.0.1:7171"},
-		Nodes:      []string{"127.0.0.1:7172", "127.0.0.1:7173", "127.0.0.1:7171"},
-		Zones:      map[string][]string{"zone": []string{"127.0.0.1:7172"}},
+		HTTPServer: &bm.ServerConfig{Addr: "127.0.0.1:7171"},
+		Nodes:      []string{"127.0.0.1:7172"},
 		Env: &dc.Env{
 			Zone:      "sh001",
 			DeployEnv: "pre",
@@ -41,9 +42,7 @@ func TestReplicate(t *testing.T) {
 		i := model.NewInstance(reg)
 		nodes := NewNodes(config)
 		nodes.nodes[0].client.SetTransport(gock.DefaultTransport)
-		nodes.nodes[1].client.SetTransport(gock.DefaultTransport)
 		httpMock("POST", "http://127.0.0.1:7172/discovery/register").Reply(200).JSON(`{"code":0}`)
-		httpMock("POST", "http://127.0.0.1:7173/discovery/register").Reply(200).JSON(`{"code":0}`)
 		err := nodes.Replicate(context.TODO(), model.Register, i, false)
 		So(err, ShouldBeNil)
 		err = nodes.Replicate(context.TODO(), model.Renew, i, false)
@@ -52,17 +51,68 @@ func TestReplicate(t *testing.T) {
 		So(err, ShouldBeNil)
 	})
 }
+func match(h *http.Request, mock *gock.Request) (ok bool, err error) {
+	ok = true
+	err = nil
+	var arg = new(model.ArgSet)
+	err = binding.Form.Bind(h, arg)
+
+	if h.URL.Path == "/discovery/set" {
+		if err != nil {
+			mock.Reply(200).JSON(`{"ts":1514341945,"code":-400}`)
+			return
+		}
+		if len(arg.Hostname) != len(arg.Status) || len(arg.Hostname) != len(arg.Metadata) {
+			mock.Reply(200).JSON(`{"ts":1514341945,"code":-400}`)
+			return
+		}
+		mock.Reply(200).JSON(`{"ts":1514341945,"code":0}`)
+	}
+	return
+}
+
+func TestReplicateSet(t *testing.T) {
+	Convey("test replicate set", t, func(c C) {
+		nodes := NewNodes(newConfig())
+		nodes.nodes[0].client.SetTransport(gock.DefaultTransport)
+		httpMock("POST", "http://127.0.0.1:7172/discovery/set").AddMatcher(match)
+		set := &model.ArgSet{
+			Region:   "shsb",
+			Zone:     "sh001",
+			Env:      "pre",
+			AppID:    "main.arch.account-service",
+			Hostname: []string{"test1", "test2"},
+			Status:   []int64{1, 1},
+			Metadata: []string{`{"aa":1,"bb:2"}`, `{"aa":1,"bb:3"}`},
+		}
+		err := nodes.ReplicateSet(context.TODO(), set, false)
+		c.So(err, ShouldBeNil)
+		set = &model.ArgSet{
+			Region:   "shsb",
+			Zone:     "sh001",
+			Env:      "pre",
+			AppID:    "main.arch.account-service",
+			Hostname: []string{"test1", "test2"},
+			Status:   []int64{1, 1},
+			Metadata: []string{`{"aa":1,"bb:2"}`},
+		}
+		err = nodes.ReplicateSet(context.TODO(), set, false)
+		c.So(err, ShouldNotBeNil)
+	})
+}
 
 func TestNodes(t *testing.T) {
 	Convey("test nodes", t, func() {
 		nodes := NewNodes(config)
 		res := nodes.Nodes()
-		So(len(res), ShouldResemble, 3)
+		So(len(res), ShouldResemble, 1)
 	})
 	Convey("test all nodes", t, func() {
-		nodes := NewNodes(config)
+		cfg := newConfig()
+		cfg.Zones = map[string][]string{"zone": []string{"127.0.0.1:7172"}}
+		nodes := NewNodes(cfg)
 		res := nodes.AllNodes()
-		So(len(res), ShouldResemble, 4)
+		So(len(res), ShouldResemble, 2)
 	})
 }
 
@@ -81,11 +131,11 @@ func TestUp(t *testing.T) {
 func TestCall(t *testing.T) {
 	Convey("test call", t, func() {
 		var res *model.Instance
-		node := newNode(config, "127.0.0.1:7172")
+		node := newNode(newConfig(), "127.0.0.1:7173")
 		node.client.SetTransport(gock.DefaultTransport)
-		httpMock("POST", "http://127.0.0.1:7172/discovery/register").Reply(200).JSON(`{"ts":1514341945,"code":-409,"data":{"region":"shsb","zone":"fuck","appid":"main.arch.account-service","env":"pre","hostname":"cs4sq","http":"","rpc":"0.0.0.0:18888","weight":2}}`)
+		httpMock("POST", "http://127.0.0.1:7174/discovery/register").Reply(200).JSON(`{"ts":1514341945,"code":-409,"data":{"region":"shsb","zone":"fuck","appid":"main.arch.account-service","env":"pre","hostname":"cs4sq","http":"","rpc":"0.0.0.0:18888","weight":2}}`)
 		i := model.NewInstance(reg)
-		err := node.call(context.TODO(), model.Register, i, "http://127.0.0.1:7172/discovery/register", &res)
+		err := node.call(context.TODO(), model.Register, i, "http://127.0.0.1:7174/discovery/register", &res)
 		So(err, ShouldResemble, ecode.Conflict)
 		So(res.AppID, ShouldResemble, "main.arch.account-service")
 	})